refactor: 디테일, 글쓰기 페이지 모듈 index,ts 정리
