+++ src/components/course/index.ts
@@ -1,0 +1,17 @@
+import Comment from "./Comment";
+import CommentInput from "./CommentInput";
+import CourseDesc from "./CourseDesc";
+import CourseHashTag from "./CourseHashTag";
+import CourseTitle from "./CourseTitle";
+import DeleteModal from "./DeleteModal";
+import LikeBtn from "./LikeBtn";
+
+export {
+  Comment,
+  CommentInput,
+  CourseDesc,
+  CourseHashTag,
+  CourseTitle,
+  DeleteModal,
+  LikeBtn,
+};
