+++ src/components/post/index.ts
@@ -1,0 +1,7 @@
+import CourseLine from "./CourseLine";
+import PostBtn from "./PostBtn";
+import PostHashTag from "./PostHashTag";
+import PostSearchModal from "./PostSearchModal";
+import PostTitle from "./PostTitle";
+
+export { CourseLine, PostBtn, PostHashTag, PostSearchModal, PostTitle };
